task: move perf.{Server,Client}Task to task.{Server,Client}Task
The perf.py module is a bit unmotivated. These base classes fit well
beside the Task and PluginTask in the task module. Move.
